导入文件的代码done，并修复ai瞎回答导致图片下标导致溢出的bug

+++ utils/PPTParser.py
@@ -210,7 +210,7 @@ class PPTParser:
                 if "image" in key:
                     image_url_content = ""
                     for image_index in result[key]:
-                        if flag[image_index - 1] == 1:
+                        if image_index > len(image_names) or flag[image_index - 1] == 1:
                             continue
                         url = image_names[image_index - 1]
                         url = self.image_dir + "/" + url
@@ -223,15 +223,16 @@ class PPTParser:
             document = Document(**result,
                                 is_vectorized=0)
 
-            for i in range(len(image_urls)):
-                if flag[i] == 0:
-                    os.remove(image_urls[i])
-                    dir_name, filename = os.path.split(image_urls[i])
-                    name, ext = os.path.splitext(filename)
-                    new_path = f"{name}_compressed.{ext}"
-                    new_path = os.path.join(dir_name, new_path)
-                    if os.path.exists(new_path):
-                        os.remove(new_path)
+            if len(image_urls) > 0:
+                for i in range(len(image_urls)):
+                    if flag[i] == 0:
+                        os.remove(image_urls[i])
+                        dir_name, filename = os.path.split(image_urls[i])
+                        name, ext = os.path.splitext(filename)
+                        new_path = f"{name}_compressed.{ext}"
+                        new_path = os.path.join(dir_name, new_path)
+                        if os.path.exists(new_path):
+                            os.remove(new_path)
 
             return document
